web: replace xml table tanstack renderer with shadcn table
